+++ integration/app/src/examples/article/article-entities.state.ts
@@ -1,7 +1,7 @@
+import { createEntityCollections } from '@angular-ru/common/entity';
 import { Injectable } from '@angular/core';
 import { Persistence, StateRepository } from '@ngxs-labs/data/decorators';
 import { NgxsDataEntityCollectionsRepository } from '@ngxs-labs/data/repositories';
-import { createEntityCollections } from '@ngxs-labs/data/utils';
 import { State } from '@ngxs/store';
 
 import { Article } from './article.interface';