+++ lib/internals/src/decorators/validate-computed-method.ts
@@ -1,10 +1,9 @@
+import { isGetter } from '@angular-ru/common/object';
 import { Any } from '@angular-ru/common/typings';
 import { NGXS_DATA_EXCEPTIONS } from '@ngxs-labs/data/tokens';
 
-import { isGetter } from '../utils/common/is-getter';
-
 export function validateComputedMethod(target: Any, name: string | symbol): void {
-    const notGetter: boolean = !isGetter(target, name);
+    const notGetter: boolean = !isGetter(target, name?.toString());
     if (notGetter) {
         throw new Error(
             NGXS_DATA_EXCEPTIONS.NGXS_COMPUTED_DECORATOR +