+++ lib/internals/src/utils/repository/ensure-snapshot.ts
@@ -1,7 +1,6 @@
+import { deepFreeze } from '@angular-ru/common/object';
 import { isDevMode } from '@angular/core';
 
-import { ngxsDeepFreeze } from '../common/freeze';
-
 export function ensureSnapshot<T>(state: T): T {
-    return isDevMode() ? ngxsDeepFreeze(state) : state;
+    return isDevMode() ? deepFreeze(state) : state;
 }