+++ lib/storage/src/utils/ensure-serialize-data.ts
@@ -1,4 +1,4 @@
-import { isNotNil } from '@ngxs-labs/data/internals';
+import { isNotNil } from '@angular-ru/common/utils';
 import { PersistenceProvider, STORAGE_DECODE_TYPE, StorageData } from '@ngxs-labs/data/typings';
 
 export function ensureSerializeData<T>(data: T | null, provider: PersistenceProvider): StorageData<T> {