+++ lib/testing/src/platform/ngxs-data-testing.module.ts
@@ -2,7 +2,6 @@ import { ApplicationRef, destroyPlatform, ModuleWithProviders, NgModule, Type } 
 import { TestBed } from '@angular/core/testing';
 import { NgxsDataPluginModule } from '@ngxs-labs/data';
 import { NGXS_DATA_STORAGE_CONTAINER, NGXS_DATA_STORAGE_EXTENSION } from '@ngxs-labs/data/storage';
-import { NgxsDataUtilsModule } from '@ngxs-labs/data/utils';
 import { NgxsModule } from '@ngxs/store';
 import { StateClass } from '@ngxs/store/internals';
 
@@ -13,8 +12,7 @@ import { NgxsAppMockModule } from './ngxs-app-mock.module';
 type NgxsDataTestingModuleProviders = [
     Type<NgxsAppMockModule>,
     ModuleWithProviders<NgxsModule>,
-    ModuleWithProviders<NgxsDataPluginModule>,
-    Type<NgxsDataUtilsModule>
+    ModuleWithProviders<NgxsDataPluginModule>
 ];
 
 @NgModule({
@@ -26,8 +24,7 @@ export class NgxsDataTestingModule {
         return [
             NgxsAppMockModule,
             NgxsModule.forRoot(states, { developmentMode: true, selectorOptions: { suppressErrors: false } }),
-            NgxsDataPluginModule.forRoot([NGXS_DATA_STORAGE_EXTENSION, NGXS_DATA_STORAGE_CONTAINER]),
-            NgxsDataUtilsModule
+            NgxsDataPluginModule.forRoot([NGXS_DATA_STORAGE_EXTENSION, NGXS_DATA_STORAGE_CONTAINER])
         ];
     }