+++ lib/typings/src/entity/entity-update.ts
@@ -1,4 +1,0 @@
-export interface EntityUpdate<V, K> {
-    id: K;
-    changes: Partial<V>;
-}