+++ lib/typings/src/entity/primary-key.ts
@@ -1,3 +1,0 @@
-export const enum PrimaryKey {
-    ID = 'id'
-}