+++ lib/utils/src/public_api.ts
@@ -1,5 +1,0 @@
-export { createEntityCollections } from './entity/create-entity-collections';
-export { entitySortByAsc } from './entity/entity-sort-by-asc';
-export { entitySortByDesc } from './entity/entity-sort-by-desc';
-export { NgxsDataMutablePipe } from './mutable/ngxs-data-mutable.pipe';
-export { NgxsDataUtilsModule } from './ngxs-data-utils.module';