+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # To become 4.0.0
 
+-   Feature: support `@ngxs/store@3.7.0`
+
 ### ⚠ BREAKING CHANGES
 
 -   Moved `Immutable, Mutable, Any` interfaces to `@angular-ru/common/typings`:
@@ -7,6 +9,26 @@
 ```ts
 import { Immutable, Mutable, Any } from '@angular-ru/common/typings';
 ```
+
+-   Moved `createEntityCollections` to `@angular-ru/common/entity`:
+
+```ts
+import { createEntityCollections } from '@angular-ru/common/entity';
+
+@StateRepository()
+@State({
+    name: 'students',
+    defaults: createEntityCollections()
+})
+@Injectable()
+class StudentEntitiesState extends NgxsDataEntityCollectionsRepository<StudentEntity, string> {
+    public selectId(entity: StudentEntity): string {
+        return `${entity.groupId}_${entity.batchId}`;
+    }
+}
+```
+
+-   Removed `NgxsDataMutablePipe, NgxsDataUtilsModule`
 
 # 3.0.6 2020-06-25